+++ services/project-creator.js
@@ -2,7 +2,8 @@ const chalk = require('chalk');
 const Authenticator = require('./authenticator');
 const api = require('./api');
 const KeyGenerator = require('./key-generator');
-const logger = require('./logger');
+const { terminate } = require('../utils/terminator');
+const { ERROR_UNEXPECTED } = require('../utils/messages');
 
 function ProjectCreator() {
   const auth = new Authenticator();
@@ -18,15 +19,18 @@ function ProjectCreator() {
         authSecret: await new KeyGenerator().generate(),
       };
     } catch (error) {
+      let message;
       if (error.message === 'Unauthorized') {
-        logger.error(`Your session has expired. Please, relogin with the command ${chalk.cyan('lumber run lumber-forestadmin:login')}.`);
+        message = `Your session has expired. Please log back in with the command ${chalk.cyan('lumber login')}.`;
       } else if (error.message === 'Conflict') {
-        logger.error('A project with this name already exists. Please choose another name.');
+        message = 'A project with this name already exists. Please choose another name.';
       } else {
-        logger.error(`An unexpected error occured. Please create a Github issue with following error: ${chalk.red(error)}`);
+        message = `${ERROR_UNEXPECTED} ${chalk.red(error)}`;
       }
 
-      return process.exit(1);
+      return terminate(1, {
+        logs: [message],
+      });
     }
   };
 }